renaming xml, and dmn references to kogito
