+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-common/src/main/resources/org/kie/workbench/common/dmn/webapp/kogito/common/DMNWebappKogitoCommon.gwt.xml
@@ -32,7 +32,7 @@
   <!-- Errai, Drools etc -->
   <inherits name="org.kie.workbench.common.profile.ProfileAPI"/>
   <inherits name="org.kie.workbench.common.services.datamodel.KieWorkbenchCommonDataModelAPI"/>
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient"/>
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
   <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
   <inherits name="org.guvnor.common.services.project.GuvnorProjectClient"/>
