+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-standalone/src/main/resources/org/kie/workbench/common/dmn/showcase/DMNShowcase.gwt.xml
@@ -31,7 +31,7 @@
   <!-- Errai, Drools etc -->
   <inherits name="org.kie.workbench.common.profile.ProfileAPI"/>
   <inherits name="org.kie.workbench.common.services.datamodel.KieWorkbenchCommonDataModelAPI"/>
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient"/>
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
   <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
   <inherits name="org.kie.workbench.common.widgets.metadata.KieWorkbenchMetadataWidget"/>
   <inherits name="org.kie.workbench.common.widgets.configresource.KieWorkbenchConfigResourceWidget"/>
