+++ kie-wb-common-kogito/kie-wb-common-kogito-api/src/main/java/org/kie/workbench/common/kogito/api/SubmarineAPIEntryPoint.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.workbench.common.submarine.api;
+package org.kie.workbench.common.kogito.api;
 
 import org.jboss.errai.ioc.client.api.EntryPoint;
 import org.jboss.errai.ui.shared.api.annotations.Bundle;
