+++ kie-wb-common-kogito/kie-wb-common-kogito-api/src/main/resources/org/kie/workbench/common/kogito/KogitoAPI.gwt.xml
