+++ kie-wb-common-kogito/kie-wb-common-kogito-client/src/main/resources/org/kie/workbench/common/kogito/KogitoClient.gwt.xml
