+++ kie-wb-common-kogito/kie-wb-common-kogito-webapp-base/src/main/resources/org/kie/workbench/common/kogito/webapp/base/KogitoWebappBase.gwt.xml
@@ -29,7 +29,7 @@
   <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
   <inherits name="org.kie.workbench.common.profile.ProfileAPI" />
 
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient" />
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient" />
 
 
 </module>
