+++ kie-wb-common-stunner/kie-wb-common-stunner-extensions/kie-wb-common-stunner-kogito/kie-wb-common-stunner-kogito-client/src/main/resources/org/kie/workbench/common/stunner/kogito/StunnerKogitoClient.gwt.xml
@@ -38,8 +38,8 @@
   <inherits name="org.kie.workbench.common.stunner.client.StunnerWidgets"/>
   <inherits name="org.kie.workbench.common.stunner.forms.StunnerFormsClient"/>
 
-  <!-- Submarine -->
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient"/>
+  <!-- Kogito -->
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
 
   <source path="client"/>
