+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-kogito-runtime/src/main/resources/org/kie/workbench/common/stunner/kogito/KogitoBPMNEditor.gwt.xml
@@ -33,7 +33,7 @@
   <!-- Errai, Drools etc -->
   <inherits name="org.kie.workbench.common.profile.ProfileAPI"/>
   <inherits name="org.kie.workbench.common.services.datamodel.KieWorkbenchCommonDataModelAPI"/>
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient"/>
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
   <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
   <inherits name="org.guvnor.common.services.project.GuvnorProjectClient"/>
