+++ kie-wb-common-stunner/kie-wb-common-stunner-showcase/kie-wb-common-stunner-showcase-standalone/src/main/resources/org/kie/workbench/common/stunner/standalone/StunnerStandaloneShowcase.gwt.xml
@@ -33,7 +33,7 @@
   <!-- Errai, Drools etc -->
   <inherits name="org.kie.workbench.common.profile.ProfileAPI"/>
   <inherits name="org.kie.workbench.common.services.datamodel.KieWorkbenchCommonDataModelAPI"/>
-  <inherits name="org.kie.workbench.common.submarine.SubmarineClient"/>
+  <inherits name="org.kie.workbench.common.kogito.KogitoClient"/>
   <inherits name="org.kie.workbench.common.widgets.KieWorkbenchWidgetsCommon"/>
   <inherits name="org.guvnor.common.services.project.GuvnorProjectClient"/>
